+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 2.6.0-beta.x
 
-- Add `signatureVerify (message: Uint8Array | string, signature: Uint8Array | string, addressOrublicKey: Uint8Array | string): VerifyResult` to verify any signature, regardless of type. Returns validity & detected crypto type.
+- Add `signatureVerify (message: Uint8Array | string, signature: Uint8Array | string, addressOrPublicKey: Uint8Array | string): VerifyResult` to verify any signature, regardless of type. Returns validity & detected crypto type.
 
 # 2.5.1 Feb 22, 2020